grabbed changes from exemplar staging, and cleaned up docker ps command

+++ bin/remove-sample.sh
@@ -1,16 +1,27 @@
 #! /bin/bash
 
-set -eou pipefail
+if [ -f ./manage.py ]; then
+    echo "removing manage.py"
+    rm -rf manage.py
+fi
 
-# Grab environment variables
-. ./.env
+if [ -d ./dead_songs ]; then
+    echo "removing django project folder"
+    rm -rf dead_songs
+fi
 
-if [ -f manage.py ]; then rm manage.py; fi
-rm -rf dead_songs
-rm -rf api
+if [ -d ./api ]; then
+    echo "removing api app folder"
+    rm -rf api
+fi
 
-if [ -f ./Backups/dead_songs.sql ]; then rm ./Backups/dead_songs.sql; fi
+if [ -f ./Backups/dead_songs.sql ]; then
+    echo "removing sample db backup"
+    rm ./Backups/dead_songs.sql
+fi
 
-# remove sample database
-docker-compose -f development-docker-compose.yml run api_development /bin/bash -c "export PGPASSWORD=${DEVELOPMENT_POSTGRES_PASSWORD}; psql -h "$DEVELOPMENT_POSTGRES_HOST" -U "$DEVELOPMENT_POSTGRES_USER" -c 'DROP DATABASE IF EXISTS dead_songs'; unset PGPASSWORD"
-
+# should we run docker-compose down if there is an api_development child?
+if [ "$(docker ps -a -q -f ancestor=db_development)" ]; then
+  echo "removing containers"
+  docker-compose -f development-docker-compose.yml down
+fi
